ft: extract functions metadata to separate file
