+++ lib/mix_unused/exports.ex
@@ -1,18 +1,7 @@
 defmodule MixUnused.Exports do
   @moduledoc false
 
-  defmodule Meta do
-    @moduledoc false
-
-    @type t() :: %__MODULE__{
-            signature: String.t(),
-            file: String.t(),
-            line: non_neg_integer(),
-            doc_meta: map()
-          }
-
-    defstruct signature: nil, file: "nofile", line: 1, doc_meta: %{}
-  end
+  alias MixUnused.Meta
 
   @type t() :: %{mfa() => Meta.t()} | [{mfa(), Meta.t()}]
