+++ test/mix_unused/analyzers/private_test.exs
@@ -1,7 +1,7 @@
 defmodule MixUnused.Analyzers.PrivateTest do
   use ExUnit.Case, async: true
 
-  alias MixUnused.Exports.Meta
+  alias MixUnused.Meta
 
   @subject MixUnused.Analyzers.Private
